+++ server/connector/tsq_term.cpp
@@ -1,0 +1,109 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2026 SereneDB GmbH, Berlin, Germany
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is SereneDB GmbH, Berlin, Germany
+////////////////////////////////////////////////////////////////////////////////
+
+#include <duckdb/planner/expression/bound_cast_expression.hpp>
+#include <iresearch/analysis/token_attributes.hpp>
+#include <iresearch/search/terms_filter.hpp>
+#include <iresearch/utils/string.hpp>
+
+#include "catalog/mangling.h"
+#include "tsq_common.hpp"
+
+namespace sdb::connector {
+
+Result SetupTermFilter(irs::ByTerm& filter, std::string& field_name,
+                       const SearchColumnInfo& column_info,
+                       const duckdb::Value& value);
+
+Result BuildFtsTerm(irs::BooleanFilter& parent, const FilterContext& ctx,
+                    const SearchColumnInfo& column_info,
+                    const duckdb::Value& value) {
+  if (value.IsNull()) {
+    AddFilter<irs::Empty>(parent);
+    return {};
+  }
+  auto& term =
+    ctx.negated ? Negate<irs::ByTerm>(parent) : AddFilter<irs::ByTerm>(parent);
+  term.boost(ctx.boost);
+  std::string field_name;
+  MakeFieldName(column_info, field_name);
+  return SetupTermFilter(term, field_name, column_info, value);
+}
+
+Result BuildFtsTokens(irs::BooleanFilter& parent, const FilterContext& ctx,
+                      const SearchColumnInfo& column_info,
+                      std::string_view text, bool require_all) {
+  if (column_info.logical_type.id() != duckdb::LogicalTypeId::VARCHAR) {
+    return BuildFtsTerm(parent, ctx, column_info,
+                        duckdb::Value(std::string{text}));
+  }
+  auto& analyzer = ctx.tokenizer;
+  std::vector<irs::bstring> tokens;
+  if (!analyzer.reset(text)) {
+    return {ERROR_BAD_PARAMETER, "Failed to analyse '", text, "'"};
+  }
+  const auto* tok_attr = irs::get<irs::TermAttr>(analyzer);
+  while (analyzer.next()) {
+    tokens.emplace_back(tok_attr->value.begin(), tok_attr->value.end());
+  }
+
+  std::string field_name;
+  MakeFieldName(column_info, field_name);
+  search::mangling::MangleString(field_name);
+
+  if (tokens.empty()) {
+    AddFilter<irs::Empty>(parent);
+    return {};
+  }
+  if (tokens.size() == 1) {
+    auto& term = ctx.negated ? Negate<irs::ByTerm>(parent)
+                             : AddFilter<irs::ByTerm>(parent);
+    term.boost(ctx.boost);
+    *term.mutable_field() = field_name;
+    term.mutable_options()->term.assign(tokens[0]);
+    return {};
+  }
+  // Multi-token: ByTerms with min_match=1 (OR) or N (AND).
+  auto& terms = ctx.negated ? Negate<irs::ByTerms>(parent)
+                            : AddFilter<irs::ByTerms>(parent);
+  terms.boost(ctx.boost);
+  *terms.mutable_field() = field_name;
+  auto& opts = *terms.mutable_options();
+  opts.min_match = require_all ? tokens.size() : 1;
+  for (auto& t : tokens) {
+    opts.terms.emplace(std::move(t));
+  }
+  return {};
+}
+Result FromTerm(irs::BooleanFilter& parent, const FilterContext& ctx,
+                const SearchColumnInfo& column_info,
+                const duckdb::BoundFunctionExpression& func) {
+  if (func.children.size() != 1) {
+    return {ERROR_BAD_PARAMETER, "TERM expects 1 argument (text), got ",
+            func.children.size()};
+  }
+  std::string text;
+  if (auto r = GetVarcharArg(*func.children[0], "TERM text", text); !r.ok()) {
+    return r;
+  }
+  return BuildFtsTerm(parent, ctx, column_info, duckdb::Value(text));
+}
+
+}  // namespace sdb::connector
